Refactor resource management to BaseServer
Eliminated the standalone ResourceManager and integrated resource handling directly into BaseServer. Adjusted relevant methods and tests to reflect the new structure. This simplifies the architecture and centralizes resource operations within BaseServer.
