+++ mcp/server_base_test.go
@@ -368,3 +368,119 @@ func TestValidatePrompt(t *testing.T) {
 		assert.Contains(t, err.Error(), "prompt must have at least one message")
 	})
 }
+
+func TestListResources(t *testing.T) {
+	baseServer, _ := NewBaseServer(
+		UseLogger(log.New(os.Stderr, "[MCP SSEServer] ", log.LstdFlags|log.Lmsgprefix)),
+	)
+
+	// Add resources in non-sequential order
+	resources := []Resource{
+		{URI: "d_res", Name: "d_res", MimeType: "text/plain"},
+		{URI: "a_res", Name: "a_res", MimeType: "text/plain"},
+		{URI: "c_res", Name: "c_res", MimeType: "text/plain"},
+		{URI: "b_res", Name: "b_res", MimeType: "text/plain"},
+	}
+
+	err := baseServer.AddResources(resources...)
+	assert.NoError(t, err)
+
+	t.Run("list_with_cursor", func(t *testing.T) {
+		// First page
+		result := baseServer.ListResources("", 2)
+		assert.Len(t, result.Resources, 2, "First page should have 2 resources")
+		assert.Equal(t, "a_res", result.Resources[0].URI)
+		assert.Equal(t, "b_res", result.Resources[1].URI)
+		assert.Equal(t, "b_res", result.NextCursor)
+
+		// Second page
+		result = baseServer.ListResources(result.NextCursor, 2)
+		assert.Len(t, result.Resources, 2, "Second page should have 2 resources")
+		assert.Equal(t, "c_res", result.Resources[0].URI)
+		assert.Equal(t, "d_res", result.Resources[1].URI)
+		assert.Empty(t, result.NextCursor)
+	})
+}
+
+func TestReadResource(t *testing.T) {
+
+	textResource := Resource{
+		URI:         "file://file.txt",
+		Name:        "file.txt",
+		MimeType:    "text/plain",
+		TextContent: "Hello, World!",
+	}
+
+	binaryResource := Resource{
+		URI:         "file://file.bin",
+		Name:        "file.bin",
+		MimeType:    "application/octet-stream",
+		TextContent: "binary content",
+	}
+
+	baseServer, _ := NewBaseServer(
+		UseLogger(log.New(os.Stderr, "[MCP SSEServer] ", log.LstdFlags|log.Lmsgprefix)),
+	)
+	err := baseServer.AddResources(textResource, binaryResource)
+	assert.NoError(t, err)
+
+	tests := []struct {
+		name        string
+		params      ReadResourceParams
+		expectError bool
+		checkText   bool
+	}{
+		{
+			name:        "read text resource",
+			params:      ReadResourceParams{URI: "file://file.txt"},
+			expectError: false,
+			checkText:   true,
+		},
+		{
+			name:        "read binary resource",
+			params:      ReadResourceParams{URI: "file://file.bin"},
+			expectError: false,
+			checkText:   false,
+		},
+		{
+			name:        "read non-existent resource",
+			params:      ReadResourceParams{URI: "file://nonexistent.txt"},
+			expectError: true,
+			checkText:   false,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			result, err := baseServer.ReadResource(tt.params)
+
+			if tt.expectError && err == nil {
+				t.Error("expected error but got nil")
+				return
+			}
+
+			if !tt.expectError {
+				if err != nil {
+					t.Errorf("unexpected error: %v", err)
+					return
+				}
+
+				if len(result.Contents) != 1 {
+					t.Errorf("expected 1 content, got %d", len(result.Contents))
+					return
+				}
+
+				content := result.Contents[0]
+				if tt.checkText {
+					if content.Text == "" {
+						t.Error("expected non-empty text content")
+					}
+				} else {
+					if content.Blob == "" {
+						t.Error("expected non-empty blob content")
+					}
+				}
+			}
+		})
+	}
+}
